requirements.txt
librosa==0.10.2.post1
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
scipy==1.15.2
sounddevice==0.5.1
streamlit==1.43.2
tensorflow_cpu==2.19.0
